feat(pow): bit-level granular PoW difficulty (Phase 1 — JS solver) (#2696)

+++ packages/types/src/decisionMachine/index.ts
@@ -234,5 +234,5 @@ export const RoutingMachineOutputSchema = z.object({
 		z.literal(CaptchaType.puzzle),
 	]),
 	solvedImagesCount: z.number().int().positive().optional(),
-	powDifficulty: z.number().int().positive().optional(),
+	powDifficulty: z.number().positive().optional(),
 });
